requirements.txt
asgiref==3.5.2
Django==3.2.10
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
